Move methods to utils
