+++ server/src/core/api-rest/api-rest-query-builder/api-rest-query-builder.factory.spec.ts
@@ -39,47 +39,4 @@ describe('ApiRestQueryBuilderFactory', () => {
   it('should be defined', () => {
     expect(service).toBeDefined();
   });
-  describe('parsePath', () => {
-    it('should parse object from request path', () => {
-      const request: any = { path: '/rest/companies/uuid' };
-
-      expect(service.parsePath(request)).toEqual({
-        object: 'companies',
-        id: 'uuid',
-      });
-    });
-
-    it('should parse object from request path', () => {
-      const request: any = { path: '/rest/companies' };
-
-      expect(service.parsePath(request)).toEqual({
-        object: 'companies',
-        id: undefined,
-      });
-    });
-  });
-  describe('computeDepth', () => {
-    it('should compute depth from query', () => {
-      const request: any = {
-        query: { depth: '1' },
-      };
-
-      expect(service.computeDepth(request)).toEqual(1);
-    });
-
-    it('should return default depth if missing', () => {
-      const request: any = { query: {} };
-
-      expect(service.computeDepth(request)).toEqual(undefined);
-    });
-    it('should raise if wrong depth', () => {
-      const request: any = { query: { depth: '100' } };
-
-      expect(() => service.computeDepth(request)).toThrow();
-
-      request.query.depth = '0';
-
-      expect(() => service.computeDepth(request)).toThrow();
-    });
-  });
 });
