+++ server/src/core/api-rest/api-rest-query-builder/factories/input-factories/__tests__/order-by-input.factory.spec.ts
@@ -3,7 +3,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { OrderByDirection } from 'src/workspace/workspace-query-builder/interfaces/record.interface';
 
 import { OrderByInputFactory } from 'src/core/api-rest/api-rest-query-builder/factories/input-factories/order-by-input.factory';
-import { objectMetadataItem } from 'src/core/api-rest/utils/__tests__/utils';
+import { objectMetadataItem } from 'src/core/api-rest/api-rest-query-builder/utils/__tests__/utils';
 
 describe('OrderByInputFactory', () => {
   const objectMetadata = { objectMetadataItem: objectMetadataItem };
