+++ server/src/core/api-rest/api-rest-query-builder/factories/input-factories/filter-utils/add-default-conjunction.utils.ts
@@ -1,0 +1,7 @@
+export const addDefaultConjunctionIfMissing = (filterQuery: string): string => {
+  if (!(filterQuery.includes('(') && filterQuery.includes(')'))) {
+    return `and(${filterQuery})`;
+  }
+
+  return filterQuery;
+};
